Implement thread safe request cache
