+++ user-interface/src/main/resources/application.properties
@@ -202,6 +202,9 @@ qbic.communication.api.url=${API_URL:https://download.qbic.uni-tuebingen.de/swag
 terminology.service.tib.endpoint.select=${TIB_SERVICE_API_SELECT_ENDPOINT:/api/select}
 terminology.service.tib.api.url=${TIB_SERVICE_API_URL:https://api.terminology.tib.eu}
 terminology.service.tib.endpoint.search=${TIB_SERVICE_API_SEARCH_ENDPOINT:/api/search}
+# Configures the number of entries the local terminology lookup cache should store at maximum
+# before overwriting entries with low usage statistics
+terminology.service.cache.size=${TERM_CACHE_SIZE:10000}
 ###############################################################################
 ################### Message providers #########################################
 spring.messages.encoding=UTF-8
